requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
strands-agents>=1.31.0